+++ src/Chirp.Razor/Pages/PagedCheeps.cshtml
@@ -26,11 +26,11 @@
                             <!-- Follow/Unfollow button -->
                             
                             <form method="post" asp-page-handler="OnPostFollowAsync" style="display:inline;">
+                                @Html.AntiForgeryToken()
                                 <input type="hidden" name="followedId" value="@cheep.Author"/>
                                 <button type="submit" class="btn btn-@(await Model.IsFollowing(cheep.Author) ? "danger" : "primary") btn-sm">
                                     @(await Model.IsFollowing(cheep.Author) ? "Unfollow" : "Follow")
                                 </button>
-                                @Html.AntiForgeryToken()
                             </form>
                         }
                     </p>
