+++ src/Chirp.Razor/Pages/PagedCheeps.cshtml.cs
@@ -7,10 +7,10 @@ using System.Threading.Tasks;
 
 namespace Chirp.Razor.Pages
 {
-    public class CheepsModel : PageModel
-    {private ICheepService _cheepService;
+    public class CheepsModel : SharedFuncs {
+        private ICheepService _cheepService;
 
-        public CheepsModel(ICheepService cheepService)
+        public CheepsModel(ICheepService cheepService) : base(cheepService)
         {
             _cheepService = cheepService;
         }
@@ -21,56 +21,7 @@ namespace Chirp.Razor.Pages
             }
             return false;
         }
-
-        public async Task<Boolean> IsFollowing(String authorname) {
-            if (await _cheepService.IsUserFollowing(User.Identity.Name,authorname))
-            {
-                return true;
-            }
-            return false;
-        }
         
-        public async Task<IActionResult> OnPostFollowAsync(string followedId) {
-            await Task.CompletedTask;
-            if (string.IsNullOrEmpty(followedId))
-            {
-                return BadRequest("Followed ID is required.");
-            }
-
-            // Ensure the user is authenticated
-            if (User.Identity?.IsAuthenticated != true)
-            {
-                return Unauthorized();
-            }
-
-
-            Console.WriteLine("Following ID is " + followedId);
-            _cheepService.FollowAuthor(followedId, User.Identity.Name);
-            return Page(); // Refresh the current page
-        }
-        
-        public async Task<IActionResult> OnPostUnfollowAsync(string followedId)
-        {
-
-            if (string.IsNullOrEmpty(followedId))
-            {
-                return BadRequest("Followed ID is required.");
-            }
-
-            // Ensure the user is authenticated
-            if (User.Identity?.IsAuthenticated != true)
-            {
-                return Unauthorized();
-            }
-            
-            //change to unfollow
-            // _cheepService.FollowAuthor(followedId, User.Identity.Name);
-            return RedirectToPage(); // Refresh the current page
-            
-        }
-        
-        
-
         public List<CheepViewModel> Cheeps { get; set; }
         public int CurrentPage { get; set; }
         public int TotalPages { get; set; }
